+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## NEXT
 
+- removed `ensRegistryAddress` option (use ethers `Network.register()` to configure custom networks supporting ENS)
 - more permissive url validation for `iexec_input_files`
 
 ## [8.9.1] 2024-06-19